% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE.R
\name{ase}
\alias{ase}
\title{ASE calculator}
\usage{
ase(x, xhat)
}
\arguments{
\item{x}{the time series}

\item{xhat}{the saved forecast (results of forecast function)}
}
\value{
the ASE
}
\description{
ASE calculator
}
\examples{
xs <- generate(arma, 200)
fore <- fcst(arma, xs, n.ahead =20, lastn = TRUE)
ase(xs, fore)
}
