% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stationarity.R
\name{calculate_ar1_varx}
\alias{calculate_ar1_varx}
\title{Computes the Population Variance of an AR(1) Time Series}
\usage{
calculate_ar1_varx(phi, vara = 1)
}
\arguments{
\item{phi}{phi values of the time series (only 1 in this case)}

\item{vara}{variance of the noise term}
}
\value{
True Population Variance of the Time Series
}
\description{
Computes the Population Variance of an AR(1) Time Series
}
