% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stationarity.R
\name{calculate_arp_varx}
\alias{calculate_arp_varx}
\title{Computes the Population Variance of an AR(p) Time Series
Note that this can also be used for AR(1) models instead of using calculate_ar1_varx
although, it will need an extra argument 'pt'}
\usage{
calculate_arp_varx(phi, pt, vara = 1)
}
\arguments{
\item{phi}{phi values of the time series}

\item{pt}{output of plotts.true.wge}

\item{vara}{variance of the noise term}
}
\value{
True Population Variance of the Time Series
}
\description{
Computes the Population Variance of an AR(p) Time Series
Note that this can also be used for AR(1) models instead of using calculate_ar1_varx
although, it will need an extra argument 'pt'
}
