% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_calculations.R
\name{calculate_ts_gamma}
\alias{calculate_ts_gamma}
\title{Computes Gamma (Auto Covariance of a time series realization)}
\usage{
calculate_ts_gamma(x, k = 0)
}
\arguments{
\item{x}{time series realization}

\item{k}{lag for which ACV needs to be calculated (Default = 0 -> Variance of the Time Series)}
}
\value{
Auto Covaraince at lag k
}
\description{
Computes Gamma (Auto Covariance of a time series realization)
}
