% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stationarity.R
\name{calculate_ts_gamma0}
\alias{calculate_ts_gamma0}
\title{Computes Gamma0 (Variance of a time series realization)}
\usage{
calculate_ts_gamma0(x)
}
\arguments{
\item{x}{time series realization}
}
\value{
Gamma0 (Variance of the time series realization)
}
\description{
Computes Gamma0 (Variance of a time series realization)
}
