% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_calculations.R
\name{calculate_ts_mean}
\alias{calculate_ts_mean}
\title{Computes the Mean of a Time Series realization}
\usage{
calculate_ts_mean(x)
}
\arguments{
\item{x}{time series realization}
}
\value{
Mean of the Time Series realization
}
\description{
Computes the Mean of a Time Series realization
}
