% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_calculations.R
\name{calculate_ts_mean_confidence_interval}
\alias{calculate_ts_mean_confidence_interval}
\title{Computes the Confidence Interval of the Mean of a Time Series}
\usage{
calculate_ts_mean_confidence_interval(x, alpha = 0.05)
}
\arguments{
\item{x}{time series realization}

\item{alpha}{alpha value to be used for calculating the Confidence Interval}
}
\value{
Confidence Interval of the Mean of a Time Series
}
\description{
Computes the Confidence Interval of the Mean of a Time Series
}
