% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_calculations.R
\name{calculate_ts_rho}
\alias{calculate_ts_rho}
\title{Computes Rho (Auto Correlation of a time series realization)}
\usage{
calculate_ts_rho(x, k = 1)
}
\arguments{
\item{x}{time series realization}

\item{k}{lag for which Auto Correlation needs to be calculated (Default = 1)}
}
\value{
Auto Correlation at lag k
}
\description{
Computes Rho (Auto Correlation of a time series realization)
}
