% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stationarity.R
\name{calculate_ts_var_of_mean}
\alias{calculate_ts_var_of_mean}
\title{Computes the Variance of the Sampling distribution of the Mean of a Time Series}
\usage{
calculate_ts_var_of_mean(x)
}
\arguments{
\item{x}{time series realization}
}
\value{
Variance of the Sampling distribution of the Mean of the Time Series
}
\description{
Computes the Variance of the Sampling distribution of the Mean of a Time Series
}
