% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stationarity.R
\name{check_stationarity}
\alias{check_stationarity}
\title{Wrapper to check the stationarity of a Time Series}
\usage{
check_stationarity(
  data,
  title = "Time Series Plot",
  xlab = "Time",
  ylab = "Time Series Realization"
)
}
\arguments{
\item{data}{time series realization}

\item{title}{Title of the time series plot (output of function)}

\item{xlab}{x-label of the time series plot (output of the function)}

\item{ylab}{y-label of the time series plot (output of the function)}
}
\description{
Wrapper to check the stationarity of a Time Series
}
