% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_calculations.R
\name{compute_a}
\alias{compute_a}
\title{Given an AR(p,q) realization, this function estimates the white 
noise estimates using equation 6.23 (from the text book)}
\usage{
compute_a(x, phi, theta, index)
}
\arguments{
\item{x}{time series realization}

\item{phi}{phi values of the time series}

\item{theta}{theta values of the time series}

\item{index}{time value till which the white noise estimates are needed}
}
\value{
All white noise estimates till the index
}
\description{
Given an AR(p,q) realization, this function estimates the white 
noise estimates using equation 6.23 (from the text book)
}
