% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_calculations.R
\name{compute_stda}
\alias{compute_stda}
\title{Given the white noise estimates, this function computes 
the standard deviation of the white noise. Note that non-zero 
terms are removed before calcualting the variance.}
\usage{
compute_stda(all_a)
}
\arguments{
\item{all_a}{White Noise Estimates}
}
\value{
Standard Deviation of the White Noise terms
}
\description{
Given the white noise estimates, this function computes 
the standard deviation of the white noise. Note that non-zero 
terms are removed before calcualting the variance.
}
