% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_calculations.R
\name{compute_vara}
\alias{compute_vara}
\title{Given the white noise estimates, this function computes 
the variance of the white noise. Note that non-zero terms 
are removed before calcualting the variance.}
\usage{
compute_vara(all_a)
}
\arguments{
\item{all_a}{White Noise Estimates}
}
\value{
Variance of the White Noise terms
}
\description{
Given the white noise estimates, this function computes 
the variance of the white noise. Note that non-zero terms 
are removed before calcualting the variance.
}
