% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{difference}
\alias{difference}
\title{Time Series transformation}
\usage{
difference(type, x, n)
}
\arguments{
\item{type}{either aruma/seasonal or arima, type of transformation to make}

\item{x}{the time series}

\item{n}{the order of transformation}
}
\value{
the transformed time series
}
\description{
Time Series transformation
}
