% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wn_tests.R
\name{evaluate_residuals}
\alias{evaluate_residuals}
\title{Evaluate if the dats is consistent with white noise}
\usage{
evaluate_residuals(
  x,
  p = 0,
  q = 0,
  k_val = c(24, 48),
  alpha = 0.05,
  lag.max = 50,
  model_name = "My Model"
)
}
\arguments{
\item{x}{the time series}

\item{p}{the ar order (Default = 0)}

\item{q}{the ma order (Default = 0)}

\item{k_val}{a vector of k values for ljung_box test}

\item{alpha}{Significance level to be used for ljung_box tests}

\item{lag.max}{Value of lags to plot for the ACF (Default: 50)}

\item{model_name}{Model name or identifier (Default: "My Model")}
}
\value{
the results of the tests, in tidy data format
}
\description{
Evaluate if the dats is consistent with white noise
}
\examples{
library(tswge)

# Generated White Noise 
wn = gen.arma.wge(n = 200, sn = 101)
evaluate_residuals(wn)

# Not White Noise
data(hadley) 
evaluate_residuals(hadley)

}
