% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_seasons.R
\name{factor.wge.season}
\alias{factor.wge.season}
\title{Prints the factor table for a pure seasonal model (p = d = q = 0)}
\usage{
factor.wge.season(s)
}
\arguments{
\item{s}{Seasonality of the model}
}
\description{
Prints the factor table for a pure seasonal model (p = d = q = 0)
}
