% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{fcst}
\alias{fcst}
\title{forecast function!}
\usage{
fcst(type, ...)
}
\arguments{
\item{type}{either arma, aruma, sigplusnoise, or other tswge model. No quotes}

\item{...}{the normal inputs to tswge}
}
\value{
a forecast
}
\description{
forecast function!
}
\examples{
fcst(arma, LakeHuron, phi = 0.2)
}
