% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gents.R
\name{generate}
\alias{generate}
\title{generator function!}
\usage{
generate(type, ...)
}
\arguments{
\item{type}{either arma, aruma, sigplusnoise, or other tswge model. No quotes}

\item{...}{the normal inputs to tswge}
}
\value{
a time series
}
\description{
generator function!
}
\examples{
generate(arma, 100, phi = 0.2, theta = 0.4)
}
