% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_multiple_realizations.R
\name{generate_multiple_realization}
\alias{generate_multiple_realization}
\title{Generate Multiple Realizations of a model
Useful for checking model appropriateness}
\usage{
generate_multiple_realization(
  x,
  phi = 0,
  theta = 0,
  d = 0,
  s = 0,
  vara = NA,
  n.realizations = 4,
  lag.max = 25,
  seed = NA,
  model_name = "Custom Model",
  return = "all"
)
}
\arguments{
\item{x}{The actual time series realization}

\item{phi}{'phi' values for the AR portion of the model (Default = 0)}

\item{theta}{'theta' values for the AR portion of the model (Default = 0)}

\item{d}{'d' value for the ARIMA portion of the model (Default = 0)}

\item{s}{seasonality 's' of the ARIMA portion of the model (Default = 0)}

\item{vara}{The white noise variance of the model}

\item{n.realizations}{The number of hypothetical realizations
to create for the model (Default = 4)}

\item{lag.max}{The maximum amount of lag data to capture (Default = 25)}

\item{seed}{The seed to use for reproducibility}

\item{model_name}{The name to be assigned to the model (Default = "Custom Model")}

\item{return}{What values to return (Default = 'all')
'all': Returns the actual along with the hypothetical realization data   
'actual': Returns only the actual realization data   
'hypothetical': Returns only the hypothetical realization data}
}
\value{
Named list with 'data' and 'results'
        'data' contains the actual and the hypothetical realizations
               in tidy data format (appropriate for plotting)
        'results' contain the ACF and Spectral Density values for the
                  actual and hypothetical realizations in tidy data format
}
\description{
Generate Multiple Realizations of a model
Useful for checking model appropriateness
}
\examples{
library(tswge)
data("sunspot.classic")
est = est.arma.wge(sunspot.classic,p=8)

r = generate_multiple_realization(x = sunspot.classic,
                                  phi = est$phi, theta = est$theta, vara = est$avar,
                                  seed = 11)
}
