% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{hush}
\alias{hush}
\title{Prevents inter 'cat' statememt from printing}
\usage{
hush(f)
}
\arguments{
\item{f}{Function to silence}
}
\description{
Prevents inter 'cat' statememt from printing
}
\examples{
hush(ljung_box(1:100, 2,4))
}
