% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wn_tests.R
\name{ljung_box}
\alias{ljung_box}
\title{ljung box test for white noise}
\usage{
ljung_box(
  x,
  p = 0,
  q = 0,
  k_val = c(24, 48),
  model_name = "My Model",
  alpha = 0.05
)
}
\arguments{
\item{x}{the time series}

\item{p}{the ar order (Default = 0)}

\item{q}{the ma order (Default = 0)}

\item{k_val}{a vector of k values}

\item{model_name}{Model name or identifier (Default: "My Model")}

\item{alpha}{Significance level to be used for ljung_box tests}
}
\value{
the results of the tests, in tidy data format
}
\description{
ljung box test for white noise
}
\examples{
library(tswge)

# Generated White Noise 
wn = gen.arma.wge(n = 200, sn = 101)
ljung_box(wn)

# Not White Noise
data(hadley) 
ljung_box(hadley)

}
