% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cor.R
\name{model_cor}
\alias{model_cor}
\title{Model Correlated data
Interactively build a model of correlated data}
\usage{
model_cor(xs, parallel = FALSE, cl = NULL)
}
\arguments{
\item{xs}{the time series to model}

\item{parallel}{whether or not to run in parallel, defaults to FALSE}

\item{cl}{the cluster in which to run on}
}
\value{
a list of model metrics and paramaters
}
\description{
Model Correlated data
Interactively build a model of correlated data
}
\examples{
xs <- playground(400)
clust <- makeCluster(2L, type = "FORK")
model_cor(xs, parallel = TRUE. cl = clust)
}
