% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_det.R
\name{model_det}
\alias{model_det}
\title{model with deterministic signal plus noise}
\usage{
model_det(xs)
}
\arguments{
\item{xs}{the time series}
}
\value{
the model parameters and metrics, as a list
}
\description{
model with deterministic signal plus noise
}
