% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBuildMultivariateVAR.R
\name{ModelBuildMultivariateVAR}
\alias{ModelBuildMultivariateVAR}
\title{R6 class ModelBuildMultivariateVAR}
\value{
A dataframe containing the following columns
         'Model': Name of the model
         'Selection': The selection criteria used for K value (AIC or BIC)
         'Trend': The trend argument used in the VARselect and VAR functions
         'SlidingASE': Whether Sliding ASE will be used for this model
         'Init_K': The K value recommended by the VARselect function
         'Final_K': The adjusted K value to take into account the smaller batch size (only when using sliding_ase)
}
\description{
R6 class ModelBuildMultivariateVAR

R6 class ModelBuildMultivariateVAR
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelBuildMultivariateVAR$new()}}
\item \href{#method-get_data}{\code{ModelBuildMultivariateVAR$get_data()}}
\item \href{#method-get_var_interest}{\code{ModelBuildMultivariateVAR$get_var_interest()}}
\item \href{#method-get_data_var_interest}{\code{ModelBuildMultivariateVAR$get_data_var_interest()}}
\item \href{#method-set_verbose}{\code{ModelBuildMultivariateVAR$set_verbose()}}
\item \href{#method-set_alpha}{\code{ModelBuildMultivariateVAR$set_alpha()}}
\item \href{#method-summarize_build}{\code{ModelBuildMultivariateVAR$summarize_build()}}
\item \href{#method-get_recommendations}{\code{ModelBuildMultivariateVAR$get_recommendations()}}
\item \href{#method-build_recommended_models}{\code{ModelBuildMultivariateVAR$build_recommended_models()}}
\item \href{#method-get_final_models}{\code{ModelBuildMultivariateVAR$get_final_models()}}
\item \href{#method-add_models}{\code{ModelBuildMultivariateVAR$add_models()}}
\item \href{#method-remove_models}{\code{ModelBuildMultivariateVAR$remove_models()}}
\item \href{#method-keep_models}{\code{ModelBuildMultivariateVAR$keep_models()}}
\item \href{#method-clone}{\code{ModelBuildMultivariateVAR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object to compare several Univatiate Time Series Models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$new(
  data = NA,
  var_interest = NA,
  mdl_list,
  alpha = 0.05,
  verbose = 0,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataframe containing the time series realizations (data should not contain time index)}

\item{\code{var_interest}}{The output variable of interest (dependent variable)}

\item{\code{mdl_list}}{A names list of all models (see format below)}

\item{\code{alpha}}{Significance level to use for filtering of variables from the recommendations}

\item{\code{verbose}}{How much to print during the model building and other processes (Default = 0)}

\item{\code{...}}{Additional parameers to feed to VARSelect (if applicable) and VAR --> Most notably "exogen"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ModelCompareMultivariateVAR` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Returns the time series realization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Time Series Realization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_var_interest}{}}}
\subsection{Method \code{get_var_interest()}}{
Returns the dependent variable name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$get_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_var_interest}{}}}
\subsection{Method \code{get_data_var_interest()}}{
Returns the dependent variable data only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$get_data_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable data only
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_verbose"></a>}}
\if{latex}{\out{\hypertarget{method-set_verbose}{}}}
\subsection{Method \code{set_verbose()}}{
Adjust the verbosity level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$set_verbose(verbose = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{0 = Minimal Printing only (usualy limited to step being performed)
1 = Basic printing of model builds, etc.
2 = Reserved for debugging mode. May slow down the run due to excessive printing, especially when using batches}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_alpha"></a>}}
\if{latex}{\out{\hypertarget{method-set_alpha}{}}}
\subsection{Method \code{set_alpha()}}{
Set the significance level to use for filtering of variables from the recommendations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$set_alpha(alpha = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Significance level to use (Default = 0.05)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize_build"></a>}}
\if{latex}{\out{\hypertarget{method-summarize_build}{}}}
\subsection{Method \code{summarize_build()}}{
Returns the VAR model Build Summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$summarize_build()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_recommendations"></a>}}
\if{latex}{\out{\hypertarget{method-get_recommendations}{}}}
\subsection{Method \code{get_recommendations()}}{
Returns a dataframe with recommended variables to use 
for each VAR model along with its corresponding lag value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$get_recommendations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame with the recommendations
(1) Number of significant variables
(2) The names of the significant variables to use
(3) Lag value to use for the model
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_recommended_models"></a>}}
\if{latex}{\out{\hypertarget{method-build_recommended_models}{}}}
\subsection{Method \code{build_recommended_models()}}{
Builds the models with the recommended lags and variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$build_recommended_models()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_final_models"></a>}}
\if{latex}{\out{\hypertarget{method-get_final_models}{}}}
\subsection{Method \code{get_final_models()}}{
Returns a final models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$get_final_models(subset = "a", mdl_names = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{The subset of models to get.
'a': All models (Default)
'u': Only User Defined Models
'r': Only the recommended models}

\item{\code{mdl_names}}{Vector of model names to get. This honors the subset variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of models
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_models"></a>}}
\if{latex}{\out{\hypertarget{method-add_models}{}}}
\subsection{Method \code{add_models()}}{
Add models to the existing object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$add_models(mdl_list, alpha = NA, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mdl_list}}{The list of new models to add}

\item{\code{alpha}}{Significance level to use for filtering of variables from the recommendations (Default = 0.05)}

\item{\code{...}}{Additional parameers to feed to VARSelect (if applicable) and VAR --> Most notably "exogen"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_models"></a>}}
\if{latex}{\out{\hypertarget{method-remove_models}{}}}
\subsection{Method \code{remove_models()}}{
Remove models from the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$remove_models(mdl_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mdl_names}}{A vector of the model names to remove.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keep_models"></a>}}
\if{latex}{\out{\hypertarget{method-keep_models}{}}}
\subsection{Method \code{keep_models()}}{
Keep only the provided models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$keep_models(mdl_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mdl_names}}{A vector of the model names to keep.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildMultivariateVAR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
