% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBuildNNforCaret.R
\name{ModelBuildNNforCaret}
\alias{ModelBuildNNforCaret}
\title{R6 class ModelBuildNNforCaret}
\value{
A dataframe containing the following columns
         'Model': Name of the model
         'Selection': The selection criteria used for K value (AIC or BIC)
         'Trend': The trend argument used in the VARselect and VAR functions
         'SlidingASE': Whether Sliding ASE will be used for this model
         'Init_K': The K value recommended by the VARselect function
         'Final_K': The adjusted K value to take into account the smaller batch size (only when using sliding_ase)
}
\description{
R6 class ModelBuildNNforCaret

R6 class ModelBuildNNforCaret
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelBuildNNforCaret$new()}}
\item \href{#method-get_data}{\code{ModelBuildNNforCaret$get_data()}}
\item \href{#method-get_data_var_interest}{\code{ModelBuildNNforCaret$get_data_var_interest()}}
\item \href{#method-summarize_build}{\code{ModelBuildNNforCaret$summarize_build()}}
\item \href{#method-get_final_models}{\code{ModelBuildNNforCaret$get_final_models()}}
\item \href{#method-clone}{\code{ModelBuildNNforCaret$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object to compare several Univatiate Time Series Models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildNNforCaret$new(
  data = NA,
  var_interest = NA,
  m = NA,
  search = "grid",
  grid = NA,
  tuneLength = NA,
  batch_size = NA,
  h = NA,
  parallel = TRUE,
  seed = 1,
  verbose = 0,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataframe containing the time series realizations (data should not contain time index)}

\item{\code{var_interest}}{The output variable of interest (dependent variable)}

\item{\code{m}}{The frequency of the variable of interest}

\item{\code{search}}{Caret grid search method: 'grid' or 'random' (Default = 'grid')}

\item{\code{grid}}{If search = 'grid', what combinations of hyperparameters to use (See format in vignette).
Allowed parameters in grid are "reps", "hd", and "allow.det.season"}

\item{\code{tuneLength}}{If search = 'random', how many random combinations to try (Default = 3)}

\item{\code{batch_size}}{Batch Size to use}

\item{\code{h}}{Forecast Horizon}

\item{\code{parallel}}{Should the grid search be run in parallel or not (Default = TRUE)}

\item{\code{seed}}{The seed to use for training the the Neural Network (Default = 1)}

\item{\code{verbose}}{How much to print during the model building and other processes (Default = 0)}

\item{\code{...}}{Additional parameers to feed to nnfor::mlp function for building the model
It is highly recommended to pass the frequency of the variable of interest 'm' to get a good model
Other arguments that can be passed can be found by typing ?nnfor::mlp in the console}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ModelBuildNNforCaret` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Returns the time series realization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildNNforCaret$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Time Series Realization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_var_interest}{}}}
\subsection{Method \code{get_data_var_interest()}}{
Returns the dependent variable data only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildNNforCaret$get_data_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable data only
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize_build"></a>}}
\if{latex}{\out{\hypertarget{method-summarize_build}{}}}
\subsection{Method \code{summarize_build()}}{
Returns the VAR model Build Summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildNNforCaret$summarize_build()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_final_models"></a>}}
\if{latex}{\out{\hypertarget{method-get_final_models}{}}}
\subsection{Method \code{get_final_models()}}{
Returns a final models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildNNforCaret$get_final_models(subset = "a")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subset}}{The subset of models to get.
'a': All models (Default)
'r': Only the recommended models}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of models
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuildNNforCaret$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
