% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCombine.R
\name{ModelCombine}
\alias{ModelCombine}
\title{R6 class ModelCompareMultivariate}
\description{
R6 class ModelCompareMultivariate

R6 class ModelCompareMultivariate
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelCombine$new()}}
\item \href{#method-get_data}{\code{ModelCombine$get_data()}}
\item \href{#method-get_var_interest}{\code{ModelCombine$get_var_interest()}}
\item \href{#method-get_data_var_interest}{\code{ModelCombine$get_data_var_interest()}}
\item \href{#method-set_verbose}{\code{ModelCombine$set_verbose()}}
\item \href{#method-plot_simple_forecasts}{\code{ModelCombine$plot_simple_forecasts()}}
\item \href{#method-plot_batch_forecasts}{\code{ModelCombine$plot_batch_forecasts()}}
\item \href{#method-plot_batch_ases}{\code{ModelCombine$plot_batch_ases()}}
\item \href{#method-plot_boxplot_ases}{\code{ModelCombine$plot_boxplot_ases()}}
\item \href{#method-statistical_compare}{\code{ModelCombine$statistical_compare()}}
\item \href{#method-get_tabular_metrics}{\code{ModelCombine$get_tabular_metrics()}}
\item \href{#method-compute_simple_forecasts}{\code{ModelCombine$compute_simple_forecasts()}}
\item \href{#method-create_ensemble}{\code{ModelCombine$create_ensemble()}}
\item \href{#method-predict_ensemble}{\code{ModelCombine$predict_ensemble()}}
\item \href{#method-clone}{\code{ModelCombine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object to compare several Univatiate Time Series Models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$new(
  data = NA,
  var_interest = NA,
  uni_models = NA,
  var_models = NA,
  mlp_models = NA,
  verbose = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataframe containing the time series realizations (data should not contain time index)}

\item{\code{var_interest}}{The output variable of interest (dependent variable)}

\item{\code{uni_models}}{A vector of ModelCompareUnivariate objects}

\item{\code{var_models}}{A vector of ModelCompareMultivariateVAR objects}

\item{\code{mlp_models}}{A vector of ModelCompareNNforCaret objects (only picks the best caret model)}

\item{\code{verbose}}{How much to print during the process (Default = 0)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ModelCombine` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Returns the time series realization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Time Series Realization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_var_interest}{}}}
\subsection{Method \code{get_var_interest()}}{
Returns the dependent variable name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$get_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_var_interest}{}}}
\subsection{Method \code{get_data_var_interest()}}{
Returns the dependent variable data only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$get_data_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable data only
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_verbose"></a>}}
\if{latex}{\out{\hypertarget{method-set_verbose}{}}}
\subsection{Method \code{set_verbose()}}{
Adjust the verbosity level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$set_verbose(verbose = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{0 = Minimal Printing only (usualy limited to step being performed)
1 = Basic printing of model builds, etc.
2 = Reserved for debugging mode. May slow down the run due to excessive printing, especially when using batches}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_simple_forecasts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_simple_forecasts}{}}}
\subsection{Method \code{plot_simple_forecasts()}}{
Plots the simple forecast for each model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$plot_simple_forecasts(
  lastn = FALSE,
  newxreg = NA,
  limits = FALSE,
  zoom = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lastn}}{If TRUE, this will plot the forecasts forthe last n.ahead values of the realization (Default: FALSE)}

\item{\code{newxreg}}{The future exogenous variable values to be used for prediction.
Applicable to models that require the values of the new exogenous variables to be provided for future forecasts, e.g. nnfor::mlp()}

\item{\code{limits}}{If TRUE, this will also plot the lower and upper limits of the forecasts (Default: FALSE)}

\item{\code{zoom}}{A number indicating how much to zoom into the plot. 
For example zoom = 50 will only plot the last 50 points of the realization
Useful for cases where realizations that are long and n.ahead is small.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_batch_forecasts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_batch_forecasts}{}}}
\subsection{Method \code{plot_batch_forecasts()}}{
Plots the forecasts per batch for all models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$plot_batch_forecasts(only_sliding = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only plot the batch forecasts 
for the models that used window ASE calculations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_batch_ases"></a>}}
\if{latex}{\out{\hypertarget{method-plot_batch_ases}{}}}
\subsection{Method \code{plot_batch_ases()}}{
Plots the ASEs per batch for all models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$plot_batch_ases(only_sliding = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only plot the ASEs for
the models that used window ASE calculations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_boxplot_ases"></a>}}
\if{latex}{\out{\hypertarget{method-plot_boxplot_ases}{}}}
\subsection{Method \code{plot_boxplot_ases()}}{
Plots the boxplot of the ASE values for the models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$plot_boxplot_ases()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-statistical_compare"></a>}}
\if{latex}{\out{\hypertarget{method-statistical_compare}{}}}
\subsection{Method \code{statistical_compare()}}{
Statistically compares the ASE values of the models using 
ANOVA and Tukey Adjustment for multiple comparison
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$statistical_compare()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The results of the ANOVA test
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tabular_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-get_tabular_metrics}{}}}
\subsection{Method \code{get_tabular_metrics()}}{
Gets the metrics and results in tabular format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$get_tabular_metrics(only_sliding = FALSE, ases = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only get results for models that use a 
sliding ASE calculation method. (Default: FALSE)}

\item{\code{ases}}{If TRUE returns the ASE values for each batch. If FALSE returns the 
forecasts, and the lower and upper limits asscoiated with the forecasts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_simple_forecasts"></a>}}
\if{latex}{\out{\hypertarget{method-compute_simple_forecasts}{}}}
\subsection{Method \code{compute_simple_forecasts()}}{
Computes the simple forecasts using all the models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$compute_simple_forecasts(lastn = FALSE, newxreg = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lastn}}{If TRUE, this will get the forecasts for the last n.ahead values of the realization (Default: FALSE). 
If there is a ModelCompareNNforCaret object passed to this object, then lastn must be TRUE.}

\item{\code{newxreg}}{The future exogenous variable values to be used for prediction.
Applicable to models that require the values of the new exogenous variables to be provided for future forecasts, e.g. nnfor::mlp()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The forecasted values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_ensemble"></a>}}
\if{latex}{\out{\hypertarget{method-create_ensemble}{}}}
\subsection{Method \code{create_ensemble()}}{
Creates an ensemble model based on all the models provided
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$create_ensemble(cuts = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_ensemble"></a>}}
\if{latex}{\out{\hypertarget{method-predict_ensemble}{}}}
\subsection{Method \code{predict_ensemble()}}{
Makes a prediction based on the ensemble model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$predict_ensemble(naive = FALSE, comb = "median", newxreg = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{naive}}{If TRUE, the ensemble will be a simple mean of the prediction of all the models
If FALSE, the ensemble will use a glm model created from the batch predictions of all the models}

\item{\code{comb}}{If 'naive' = TRUE, how to combine the predictions. Allowed values are 'mean' or 'median'}

\item{\code{newxreg}}{The future exogenous variable values to be used for prediction.
Applicable to models that require the values of the new exogenous variables to be provided for future forecasts, e.g. nnfor::mlp()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The predictions from each model along with the ensemble prediction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCombine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
