% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCompareBase.R
\name{ModelCompareBase}
\alias{ModelCompareBase}
\title{R6 class ModelCompareBase}
\description{
R6 class ModelCompareBase

R6 class ModelCompareBase
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelCompareBase$new()}}
\item \href{#method-get_data}{\code{ModelCompareBase$get_data()}}
\item \href{#method-get_data_var_interest}{\code{ModelCompareBase$get_data_var_interest()}}
\item \href{#method-get_batch_size}{\code{ModelCompareBase$get_batch_size()}}
\item \href{#method-set_batch_size}{\code{ModelCompareBase$set_batch_size()}}
\item \href{#method-get_n.ahead}{\code{ModelCompareBase$get_n.ahead()}}
\item \href{#method-set_verbose}{\code{ModelCompareBase$set_verbose()}}
\item \href{#method-get_model_names}{\code{ModelCompareBase$get_model_names()}}
\item \href{#method-add_models}{\code{ModelCompareBase$add_models()}}
\item \href{#method-remove_models}{\code{ModelCompareBase$remove_models()}}
\item \href{#method-keep_models}{\code{ModelCompareBase$keep_models()}}
\item \href{#method-compute_metrics}{\code{ModelCompareBase$compute_metrics()}}
\item \href{#method-plot_simple_forecasts}{\code{ModelCompareBase$plot_simple_forecasts()}}
\item \href{#method-plot_batch_forecasts}{\code{ModelCompareBase$plot_batch_forecasts()}}
\item \href{#method-plot_batch_ases}{\code{ModelCompareBase$plot_batch_ases()}}
\item \href{#method-plot_boxplot_ases}{\code{ModelCompareBase$plot_boxplot_ases()}}
\item \href{#method-statistical_compare}{\code{ModelCompareBase$statistical_compare()}}
\item \href{#method-get_tabular_metrics}{\code{ModelCompareBase$get_tabular_metrics()}}
\item \href{#method-summarize_build}{\code{ModelCompareBase$summarize_build()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object to compare several Univatiate Time Series Models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$new(
  data = NA,
  mdl_list,
  n.ahead = NA,
  batch_size = NA,
  step_n.ahead = TRUE,
  verbose = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataframe containing the time series realizations (data should not contain time index)}

\item{\code{mdl_list}}{A named list of all models (see format below)}

\item{\code{n.ahead}}{The number of observations used to calculate ASE or forecast ahead}

\item{\code{batch_size}}{If any of the models used sliding ase method,
then this number indicates the batch size to use}

\item{\code{step_n.ahead}}{If using sliding window, should batches be incremented by n.ahead
(Default = TRUE)}

\item{\code{verbose}}{How much to print during the model building and other processes (Default = 0)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ModelCompareBase` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Returns the time series realization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Time Series Realization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_var_interest}{}}}
\subsection{Method \code{get_data_var_interest()}}{
Returns the dependent variable data only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$get_data_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable data only
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_batch_size"></a>}}
\if{latex}{\out{\hypertarget{method-get_batch_size}{}}}
\subsection{Method \code{get_batch_size()}}{
Returns the batch size value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$get_batch_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Batch Size Value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_batch_size"></a>}}
\if{latex}{\out{\hypertarget{method-set_batch_size}{}}}
\subsection{Method \code{set_batch_size()}}{
Sets the batch size value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$set_batch_size(batch_size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{batch_size}}{Batch Size Value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_n.ahead"></a>}}
\if{latex}{\out{\hypertarget{method-get_n.ahead}{}}}
\subsection{Method \code{get_n.ahead()}}{
Returns the n.ahead value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$get_n.ahead()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The n.ahead value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_verbose"></a>}}
\if{latex}{\out{\hypertarget{method-set_verbose}{}}}
\subsection{Method \code{set_verbose()}}{
Adjust the verbosity level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$set_verbose(verbose = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{0 = Minimal Printing only (usualy limited to step being performed)
1 = Basic printing of model builds, etc.
2 = Reserved for debugging mode. May slow down the run due to excessive printing, especially when using batches}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model_names"></a>}}
\if{latex}{\out{\hypertarget{method-get_model_names}{}}}
\subsection{Method \code{get_model_names()}}{
Get all the model names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$get_model_names(only_sliding = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only plot the ASEs for
the models that used window ASE calculations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of all model names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_models"></a>}}
\if{latex}{\out{\hypertarget{method-add_models}{}}}
\subsection{Method \code{add_models()}}{
Add models to the existing object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$add_models(mdl_list, batch_size = -1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mdl_list}}{The list of new models to add}

\item{\code{batch_size}}{The batch size to use if model is using sliding ASE calculations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_models"></a>}}
\if{latex}{\out{\hypertarget{method-remove_models}{}}}
\subsection{Method \code{remove_models()}}{
Remove models from the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$remove_models(mdl_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mdl_names}}{A vector of the model names to remove.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keep_models"></a>}}
\if{latex}{\out{\hypertarget{method-keep_models}{}}}
\subsection{Method \code{keep_models()}}{
Keep only the provided models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$keep_models(mdl_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mdl_names}}{A vector of the model names to keep.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-compute_metrics}{}}}
\subsection{Method \code{compute_metrics()}}{
Computes the Metrics for the models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$compute_metrics(step_n.ahead = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_n.ahead}}{If TRUE, for rolling window calculations, the step size 
will be equal to n.ahead, else it will be equal to 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_simple_forecasts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_simple_forecasts}{}}}
\subsection{Method \code{plot_simple_forecasts()}}{
Plots the simple forecast for each model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$plot_simple_forecasts(
  lastn = FALSE,
  newxreg = NA,
  limits = FALSE,
  zoom = NA,
  plot = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lastn}}{If TRUE, this will plot the forecasts forthe last n.ahead values of the realization (Default: FALSE)}

\item{\code{newxreg}}{The future exogenous variable values. Applicable to models that require the values of the new exogenous variables to be provided for future forecasts, e.g. nnfor::mlp()}

\item{\code{limits}}{If TRUE, this will also plot the lower and upper limits of the forecasts (Default: FALSE)}

\item{\code{zoom}}{A number indicating how much to zoom into the plot. 
For example zoom = 50 will only plot the last 50 points of the realization
Useful for cases where realizations that are long and n.ahead is small.}

\item{\code{plot}}{If FALSE the plots are not plotted; useful when you want to just return the data (Default = TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_batch_forecasts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_batch_forecasts}{}}}
\subsection{Method \code{plot_batch_forecasts()}}{
Plots the forecasts per batch for all models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$plot_batch_forecasts(
  only_sliding = TRUE,
  plot = TRUE,
  silent = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only plot the batch forecasts 
for the models that used window ASE calculations}

\item{\code{plot}}{If FALSE the plots are not plotted; useful when you want to just return the data (Default = TRUE)}

\item{\code{silent}}{If FALSE, any warnings are suppressed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_batch_ases"></a>}}
\if{latex}{\out{\hypertarget{method-plot_batch_ases}{}}}
\subsection{Method \code{plot_batch_ases()}}{
Plots the ASEs per batch for all models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$plot_batch_ases(
  only_sliding = TRUE,
  plot = TRUE,
  silent = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only plot the ASEs for
the models that used window ASE calculations}

\item{\code{plot}}{If FALSE the plots are not plotted; useful when you want to just return the data (Default = TRUE)}

\item{\code{silent}}{If FALSE, any warnings are suppressed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_boxplot_ases"></a>}}
\if{latex}{\out{\hypertarget{method-plot_boxplot_ases}{}}}
\subsection{Method \code{plot_boxplot_ases()}}{
Plots the boxplot of the ASE values for the models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$plot_boxplot_ases(plot = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot}}{If FALSE the plots are not plotted; useful when you want to just return the data (Default = TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-statistical_compare"></a>}}
\if{latex}{\out{\hypertarget{method-statistical_compare}{}}}
\subsection{Method \code{statistical_compare()}}{
Statistically compares the ASE values of the models using 
ANOVA and Tukey Adjustment for multiple comparison
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$statistical_compare()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The results of the ANOVA test
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tabular_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-get_tabular_metrics}{}}}
\subsection{Method \code{get_tabular_metrics()}}{
Gets the metrics and results in tabular format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$get_tabular_metrics(only_sliding = FALSE, ases = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only get results for models that use a 
sliding ASE calculation method. (Default: FALSE)}

\item{\code{ases}}{If TRUE returns the ASE values for each batch. If FALSE returns the 
forecasts, and the lower and upper limits asscoiated with the forecasts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize_build"></a>}}
\if{latex}{\out{\hypertarget{method-summarize_build}{}}}
\subsection{Method \code{summarize_build()}}{
Returns the model Build Summary (if applicable)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareBase$summarize_build()}\if{html}{\out{</div>}}
}

}
}
