% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCompareMultivaraite.R
\name{ModelCompareMultivariate}
\alias{ModelCompareMultivariate}
\title{R6 class ModelCompareMultivariate}
\description{
R6 class ModelCompareMultivariate

R6 class ModelCompareMultivariate
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelCompareMultivariate$new()}}
\item \href{#method-get_data}{\code{ModelCompareMultivariate$get_data()}}
\item \href{#method-get_var_interest}{\code{ModelCompareMultivariate$get_var_interest()}}
\item \href{#method-get_data_var_interest}{\code{ModelCompareMultivariate$get_data_var_interest()}}
\item \href{#method-get_batch_size}{\code{ModelCompareMultivariate$get_batch_size()}}
\item \href{#method-get_n.ahead}{\code{ModelCompareMultivariate$get_n.ahead()}}
\item \href{#method-add_models}{\code{ModelCompareMultivariate$add_models()}}
\item \href{#method-compute_metrics}{\code{ModelCompareMultivariate$compute_metrics()}}
\item \href{#method-get_xIC}{\code{ModelCompareMultivariate$get_xIC()}}
\item \href{#method-plot_simple_forecasts}{\code{ModelCompareMultivariate$plot_simple_forecasts()}}
\item \href{#method-plot_batch_forecasts}{\code{ModelCompareMultivariate$plot_batch_forecasts()}}
\item \href{#method-plot_batch_ases}{\code{ModelCompareMultivariate$plot_batch_ases()}}
\item \href{#method-plot_histogram_ases}{\code{ModelCompareMultivariate$plot_histogram_ases()}}
\item \href{#method-statistical_compare}{\code{ModelCompareMultivariate$statistical_compare()}}
\item \href{#method-get_tabular_metrics}{\code{ModelCompareMultivariate$get_tabular_metrics()}}
\item \href{#method-clone}{\code{ModelCompareMultivariate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object to compare several Univatiate Time Series Models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$new(
  data = NA,
  var_interest = NA,
  mdl_list,
  n.ahead = NA,
  batch_size = NA,
  step_n.ahead = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataframe containing the time series realizations (data should not contain time index)}

\item{\code{var_interest}}{The output variable of interest (dependent variable)}

\item{\code{mdl_list}}{A names list of all models (see format below)}

\item{\code{n.ahead}}{The number of observations used to calculate ASE or forecast ahead}

\item{\code{batch_size}}{If any of the models used sliding ase method,
then this number indicates the batch size to use}

\item{\code{step_n.ahead}}{If using sliding window, should batches be incremented by n.ahead
(Default = TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ModelCompareUnivariate` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Returns the time series realization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Time Series Realization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_var_interest}{}}}
\subsection{Method \code{get_var_interest()}}{
Returns the dependent variable name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$get_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_var_interest}{}}}
\subsection{Method \code{get_data_var_interest()}}{
Returns the dependent variable data only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$get_data_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable data only
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_batch_size"></a>}}
\if{latex}{\out{\hypertarget{method-get_batch_size}{}}}
\subsection{Method \code{get_batch_size()}}{
Returns the batch size value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$get_batch_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Batch Size Value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_n.ahead"></a>}}
\if{latex}{\out{\hypertarget{method-get_n.ahead}{}}}
\subsection{Method \code{get_n.ahead()}}{
Returns the n.ahead value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$get_n.ahead()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The n.ahead value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_models"></a>}}
\if{latex}{\out{\hypertarget{method-add_models}{}}}
\subsection{Method \code{add_models()}}{
Add models to the existing object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$add_models(mdl_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mdl_list}}{The list of new models to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-compute_metrics}{}}}
\subsection{Method \code{compute_metrics()}}{
Computes the Metrics for the models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$compute_metrics(step_n.ahead = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_n.ahead}}{If TRUE, for rolling window calculations, the step size 
will be equal to n.ahead, else it will be equal to 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_xIC"></a>}}
\if{latex}{\out{\hypertarget{method-get_xIC}{}}}
\subsection{Method \code{get_xIC()}}{
Returns the AIC and the BIC for the model using the entire dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$get_xIC(sort_by = "AIC")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sort_by}}{'AIC' or 'BIC'. Selects which column to sort the results by (Default: 'AIC')}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_simple_forecasts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_simple_forecasts}{}}}
\subsection{Method \code{plot_simple_forecasts()}}{
Plots the simple forecast for each model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$plot_simple_forecasts(lastn = FALSE, limits = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lastn}}{If TRUE, this will plot the forecasts forthe last n.ahead values of the realization (Default: FALSE)}

\item{\code{limits}}{If TRUE, this will also plot the lower and upper limits of the forecasts (Default: FALSE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_batch_forecasts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_batch_forecasts}{}}}
\subsection{Method \code{plot_batch_forecasts()}}{
Plots the forecasts per batch for all models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$plot_batch_forecasts(only_sliding = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only plot the batch forecasts 
for the models that used window ASE calculations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_batch_ases"></a>}}
\if{latex}{\out{\hypertarget{method-plot_batch_ases}{}}}
\subsection{Method \code{plot_batch_ases()}}{
Plots the ASEs per batch for all models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$plot_batch_ases(only_sliding = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only plot the ASEs for
the models that used window ASE calculations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_histogram_ases"></a>}}
\if{latex}{\out{\hypertarget{method-plot_histogram_ases}{}}}
\subsection{Method \code{plot_histogram_ases()}}{
Plots the histogram of the ASE values for the models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$plot_histogram_ases()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-statistical_compare"></a>}}
\if{latex}{\out{\hypertarget{method-statistical_compare}{}}}
\subsection{Method \code{statistical_compare()}}{
Statistically compares the ASE values of the models using 
ANOVA and Tukey Adjustment for multiple comparison
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$statistical_compare()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The results of the ANOVA test
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tabular_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-get_tabular_metrics}{}}}
\subsection{Method \code{get_tabular_metrics()}}{
Gets the metrics and results in tabular format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$get_tabular_metrics(only_sliding = FALSE, ases = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_sliding}}{If TRUE, this will only get results for models that use a 
sliding ASE calculation method. (Default: FALSE)}

\item{\code{ases}}{If TRUE returns the ASE values for each batch. If FALSE returns the 
forecasts, and the lower and upper limits asscoiated with the forecasts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareMultivariate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
