% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCompareNNforCaret.R
\name{ModelCompareNNforCaret}
\alias{ModelCompareNNforCaret}
\title{R6 class ModelCompareMultivariate}
\description{
R6 class ModelCompareMultivariate

R6 class ModelCompareMultivariate
}
\section{Super class}{
\code{\link[tswgewrapped:ModelCompareBase]{tswgewrapped::ModelCompareBase}} -> \code{ModelCompareNNforCaret}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelCompareNNforCaret$new()}}
\item \href{#method-get_var_interest}{\code{ModelCompareNNforCaret$get_var_interest()}}
\item \href{#method-get_data_var_interest}{\code{ModelCompareNNforCaret$get_data_var_interest()}}
\item \href{#method-get_best_model_id}{\code{ModelCompareNNforCaret$get_best_model_id()}}
\item \href{#method-summarize_build}{\code{ModelCompareNNforCaret$summarize_build()}}
\item \href{#method-clone}{\code{ModelCompareNNforCaret$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="add_models">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-add_models}{\code{tswgewrapped::ModelCompareBase$add_models()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="compute_metrics">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-compute_metrics}{\code{tswgewrapped::ModelCompareBase$compute_metrics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="get_batch_size">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-get_batch_size}{\code{tswgewrapped::ModelCompareBase$get_batch_size()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="get_data">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-get_data}{\code{tswgewrapped::ModelCompareBase$get_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="get_model_names">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-get_model_names}{\code{tswgewrapped::ModelCompareBase$get_model_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="get_n.ahead">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-get_n.ahead}{\code{tswgewrapped::ModelCompareBase$get_n.ahead()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="get_tabular_metrics">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-get_tabular_metrics}{\code{tswgewrapped::ModelCompareBase$get_tabular_metrics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="keep_models">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-keep_models}{\code{tswgewrapped::ModelCompareBase$keep_models()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="plot_batch_ases">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-plot_batch_ases}{\code{tswgewrapped::ModelCompareBase$plot_batch_ases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="plot_batch_forecasts">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-plot_batch_forecasts}{\code{tswgewrapped::ModelCompareBase$plot_batch_forecasts()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="plot_boxplot_ases">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-plot_boxplot_ases}{\code{tswgewrapped::ModelCompareBase$plot_boxplot_ases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="plot_simple_forecasts">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-plot_simple_forecasts}{\code{tswgewrapped::ModelCompareBase$plot_simple_forecasts()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="remove_models">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-remove_models}{\code{tswgewrapped::ModelCompareBase$remove_models()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="set_batch_size">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-set_batch_size}{\code{tswgewrapped::ModelCompareBase$set_batch_size()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="set_verbose">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-set_verbose}{\code{tswgewrapped::ModelCompareBase$set_verbose()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tswgewrapped" data-topic="ModelCompareBase" data-id="statistical_compare">}\href{../../tswgewrapped/html/ModelCompareBase.html#method-statistical_compare}{\code{tswgewrapped::ModelCompareBase$statistical_compare()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object to compare several Univatiate Time Series Models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareNNforCaret$new(data = NA, var_interest = NA, mdl_list, verbose = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataframe containing the time series realizations (data should not contain time index)}

\item{\code{var_interest}}{The output variable of interest (dependent variable)}

\item{\code{mdl_list}}{A single caret model (which may contain results of grid or random search)}

\item{\code{verbose}}{How much to print during the model building and other processes (Default = 0)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ModelCompareNNforCaret` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_var_interest}{}}}
\subsection{Method \code{get_var_interest()}}{
Returns the dependent variable name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareNNforCaret$get_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_var_interest}{}}}
\subsection{Method \code{get_data_var_interest()}}{
Returns the dependent variable data only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareNNforCaret$get_data_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable data only
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_best_model_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_best_model_id}{}}}
\subsection{Method \code{get_best_model_id()}}{
Returns the model ID for the best model from the search
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareNNforCaret$get_best_model_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The model ID for the best model from the search
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize_build"></a>}}
\if{latex}{\out{\hypertarget{method-summarize_build}{}}}
\subsection{Method \code{summarize_build()}}{
Not applicable for the nnfor::mlp models, since we are passing already build models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareNNforCaret$summarize_build()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareNNforCaret$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
