% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateEDA.R
\name{MultivariateEDA}
\alias{MultivariateEDA}
\title{R6 class MultivariateEDA}
\description{
R6 class MultivariateEDA

R6 class MultivariateEDA
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultivariateEDA$new()}}
\item \href{#method-get_data}{\code{MultivariateEDA$get_data()}}
\item \href{#method-get_var_interest}{\code{MultivariateEDA$get_var_interest()}}
\item \href{#method-get_var_time}{\code{MultivariateEDA$get_var_time()}}
\item \href{#method-get_data_var_interest}{\code{MultivariateEDA$get_data_var_interest()}}
\item \href{#method-set_verbose}{\code{MultivariateEDA$set_verbose()}}
\item \href{#method-plot_data}{\code{MultivariateEDA$plot_data()}}
\item \href{#method-plot_scatterplots}{\code{MultivariateEDA$plot_scatterplots()}}
\item \href{#method-plot_ccf_analysis}{\code{MultivariateEDA$plot_ccf_analysis()}}
\item \href{#method-plot_lag_plots}{\code{MultivariateEDA$plot_lag_plots()}}
\item \href{#method-clone}{\code{MultivariateEDA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object to perform EDA on a Multivariate Time Series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$new(data = NA, var_interest = NA, var_time = NA, verbose = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The dataframe (or type that can be coerced to a dataframe) containing the time series realizations}

\item{\code{var_interest}}{The output variable of interest (dependent variable)}

\item{\code{var_time}}{If the dataframe has a time column, what is the name of this column?}

\item{\code{verbose}}{How much to print during the model building and other processes (Default = 0)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MultivariateEDA` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Returns the time series realization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$get_data(time = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{NA will return the original data without the 'var_time' column
'original' will return the original data with any 'var_time' column (if applicable)
'sub' will return the original data without any 'var_time' column (if applicable)
      but with a substitute 'Time' column which is equal to the observation number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The Time Series Realization
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_var_interest}{}}}
\subsection{Method \code{get_var_interest()}}{
Returns the dependent variable name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$get_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_var_time"></a>}}
\if{latex}{\out{\hypertarget{method-get_var_time}{}}}
\subsection{Method \code{get_var_time()}}{
Returns the time  variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$get_var_time()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The time variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_var_interest"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_var_interest}{}}}
\subsection{Method \code{get_data_var_interest()}}{
Returns the dependent variable data only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$get_data_var_interest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The dependent variable data only
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_verbose"></a>}}
\if{latex}{\out{\hypertarget{method-set_verbose}{}}}
\subsection{Method \code{set_verbose()}}{
Adjust the verbosity level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$set_verbose(verbose = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{0 = Minimal Printing only (usualy limited to step being performed)
1 = Basic printing of model builds, etc.
2 = Reserved for debugging mode. May slow down the run due to excessive printing, especially when using batches}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
Plots the time series with all the dependent variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$plot_data(ncol = 1, scales = "free_y", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ncol}}{Number of columns to use to show the data}

\item{\code{scales}}{The scales argument to be passed to ggplot facet_wrap layer
(Default = 'free_y') Other appropriate options: 'fixed'}

\item{\code{...}}{Arguments to pass to facet wrap. Example "ncol = 3, scales = 'free_y'"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_scatterplots"></a>}}
\if{latex}{\out{\hypertarget{method-plot_scatterplots}{}}}
\subsection{Method \code{plot_scatterplots()}}{
Plots the scatterplots matric of all the variables in the dta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$plot_scatterplots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_ccf_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-plot_ccf_analysis}{}}}
\subsection{Method \code{plot_ccf_analysis()}}{
Plots the CCF function for the dependent variable against all independent variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$plot_ccf_analysis(lag.max = 12, negative_only = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lag.max}}{The maximum lag to evaluate}

\item{\code{negative_only}}{Whether to take max cross correlation of only negative lags for the independent variables.
Many times durign predictions, we dont have future values available for the independent variables
In such cases, we can not use positive lag values for predictions. (Default = TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe containing 
        (1) 'variable': the dependent variable name, 
        (2) 'max_ccf_index': lag at which max cross correlation occurs
        (3) 'max_ccf_value': max cross correlation value (abs)
        (4) 'max_ccf_index_adjusted': adjusted index (if negative_only is FALSE, then this will show a value capped 
             at lag = 0 for any positive lag index). User may then decide to use either this value or the 'max_ccf_index'
             dependig on if the positive lag values of the dependent variable will be available for the prediction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_lag_plots"></a>}}
\if{latex}{\out{\hypertarget{method-plot_lag_plots}{}}}
\subsection{Method \code{plot_lag_plots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$plot_lag_plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateEDA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
