% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overfit.R
\name{overfit}
\alias{overfit}
\title{overfit function
produces an overfit table}
\usage{
overfit(xs, p, type = "burg")
}
\arguments{
\item{xs}{a time series}

\item{p}{the order of overfitting}

\item{type}{the type of estimate to make, defaults to burg}
}
\value{
nothing, simply displays the factor table
}
\description{
overfit function
produces an overfit table
}
\examples{
forecast(arma, LakeHuron, phi = 0.2)
test <- generate(aruma, 400, s = 4)
overfit(test,20)
}
