% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playground.R
\name{phicheck}
\alias{phicheck}
\title{check if ar component is stationary (stolen from the R stats source code)}
\usage{
phicheck(phi)
}
\arguments{
\item{phi}{phi for AR component}
}
\description{
check if ar component is stationary (stolen from the R stats source code)
}
