% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playground.R
\name{playground}
\alias{playground}
\title{Playground: generate a random time series for practice}
\usage{
playground(n, plot = TRUE)
}
\arguments{
\item{n}{the length of the resulting time series}

\item{plot}{TRUE|FALSE Should the resulting realization be plotted?}
}
\value{
a random time series, TODO add weights
}
\description{
Playground: generate a random time series for practice
}
