% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_multiple_realizations.R
\name{plot_multiple_realizations}
\alias{plot_multiple_realizations}
\title{Plot multiple realizations of a model
Useful for checking model appropriateness}
\usage{
plot_multiple_realizations(data, results, scales = "free_y")
}
\arguments{
\item{data}{The actual and the hypothetical realizations in tidy data format}

\item{results}{The ACF and Spectral Density values for the 
actual and hypothetical realizations in tidy data format}

\item{scales}{The scales argument to be passed to ggplot facet_wrap layer
(Default = 'free_y') Other appropriate options: 'fixed'}
}
\description{
Plot multiple realizations of a model
Useful for checking model appropriateness
}
\examples{
library(tswge)
data("sunspot.classic")
est = est.arma.wge(sunspot.classic,p=8)

r = generate_multiple_realization(x = sunspot.classic,
                                  phi = est$phi, theta = est$theta, vara = est$avar,
                                  seed = 11)
plot_multiple_realizations(data = r$data, results = r$results)
}
