% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{plot_res}
\alias{plot_res}
\title{plot the residuals of a time series}
\usage{
plot_res(res)
}
\arguments{
\item{res}{the residuals of a model}
}
\value{
nothing, makes a plot
}
\description{
plot the residuals of a time series
}
