% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playground.R
\name{rarima}
\alias{rarima}
\title{radnom arima}
\usage{
rarima(n, p)
}
\arguments{
\item{n}{the length of the resulting time series}

\item{p}{TRUE|FALSE Should the resulting realization be plotted?}
}
\description{
radnom arima
}
