% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_calculations.R
\name{return_all_a_calc}
\alias{return_all_a_calc}
\title{Given an AR(p,q) realization, this function estimates the white 
noise estimates using equation 6.23 (from the text book), the 
variance and the standard deviation of the white noise}
\usage{
return_all_a_calc(x, phi, theta)
}
\arguments{
\item{x}{time series realization}

\item{phi}{phi values of the time series}

\item{theta}{theta values of the time series}

\item{index}{time value till which the white noise estimates are needed}
}
\value{
All white noise estimates, variance and standard deviation of these estimates
}
\description{
Given an AR(p,q) realization, this function estimates the white 
noise estimates using equation 6.23 (from the text book), the 
variance and the standard deviation of the white noise
}
