% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_ase.R
\name{sliding_ase}
\alias{sliding_ase}
\title{Function to calculate the sliding window ASE for a model
Supports ARMA, ARIMA, ARUMA (seasonal ARIMA) and Signal Plus Noise Models}
\usage{
sliding_ase(x, phi = 0, theta = 0, d = 0, s = 0, linear = NA,
  freq = NA, n.ahead = NA, batch_size = NA,
  time_index = "test_start", ...)
}
\arguments{
\item{x}{time series realization}

\item{phi}{phi values associated with the ARIMA model}

\item{theta}{theta values associated with the ARIMA model}

\item{d}{differencing 'd' associated with the ARIMA model}

\item{s}{seasonality 's' of the ARIMA model}

\item{linear}{(TRUE|FALSE) If using a Signal Plus Noise model, should a linear signal be used?}

\item{freq}{If using a sinusoidal signal, what is the frequency of the signal?}

\item{n.ahead}{last n.ahead data points in each batch will be used for prediction and ASE calculations}

\item{batch_size}{Window Size used}

\item{time_index}{What time index to return for each ASE. Options are data_start, data_end, test_start and test_end (Defaults to test_start)}

\item{...}{any additional arguments to be passed to the forecast functions (e.g. max.p for sigplusnoise model, lambda for ARUMA models)}
}
\value{
Named list ('ASEs' - ASE values and 'Time' - Corresponding Time stamps based on time_index argument)
}
\description{
Function to calculate the sliding window ASE for a model
Supports ARMA, ARIMA, ARUMA (seasonal ARIMA) and Signal Plus Noise Models
}
