% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_ase.R
\name{sliding_ase_var}
\alias{sliding_ase_var}
\title{Function to calculate the sliding window ASE for a model
Supports VAR Model from the vars package}
\usage{
sliding_ase_var(
  data,
  var_interest,
  k,
  trend_type = NA,
  season = NULL,
  n.ahead = NA,
  batch_size = NA,
  step_n.ahead = TRUE,
  verbose = 0,
  ...
)
}
\arguments{
\item{data}{The dataframe containing the time series realizations (data should not contain time index)}

\item{var_interest}{The output variable of interest (dependent variable)}

\item{k}{The lag value to use for the VAR model (generally determined by the VARselect function)}

\item{trend_type}{The trend type to use in VARselect and the VAR model. Refer to vars::VARselect and vars::VAR for valid options.}

\item{season}{The seasonality to use in the VAR model.}

\item{n.ahead}{last n.ahead data points in each batch will be used for prediction and ASE calculations}

\item{batch_size}{Window Size used}

\item{step_n.ahead}{Whether to step each batch by n.ahead values (Default = FALSE)}

\item{verbose}{How much to print during the model building and other processes (Default = 0)}

\item{...}{Additional arguments to pass to the VAR model}
}
\value{
Named list 
        'ASEs' - ASE values
        'time_test_start' - Time Index indicating start of test time corresponding to the ASE values
        'time_test_end' - Time Index indicating end of test time corresponding to the ASE values
        'batch_num' - Indicates the batch number for each ASE value
        'AICs' = The AIC values for the individual batches
        'BICs' = The BIC values for the individual batches
        'f' - Forecasts for each batch
        'll' - Lower Forecast Limit for each batch
        'ul' - Upper Forecast Limit for each batch
        'time.forecasts' - Time Corresponding to each forecast, upper and lower limit values
}
\description{
Function to calculate the sliding window ASE for a model
Supports VAR Model from the vars package
}
