% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_difference.R
\name{transform}
\alias{transform}
\title{Time Series transformation}
\usage{
transform(type, x, n)
}
\arguments{
\item{type}{a string or just plain text, either arima or aruma/seasonal}

\item{x}{a time series}

\item{n}{the depth to transform the time series}
}
\description{
Time Series transformation
}
\examples{
ts <- generate(arima, n = 100, d = 2)
transform(arima, ts,2)
ts2 <- generate(aruma, n = 100, s =12)
transform("seasonal", ts2, n = 12)
}
