% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tswgewrapped-package.R
\docType{package}
\name{tswgewrapped-package}
\alias{tswgewrapped}
\alias{tswgewrapped-package}
\title{tswgewrapped: Helpful wrappers for 'tswge', 'vars' and 'nnfor' time series packages}
\description{
This package provides several helpful wrappers for the already useful 'tswge', 'vars' and 'nnfor' package.
}
\author{
\strong{Maintainer}: David Josephs \email{josephsd@smu.edu}

Authors:
\itemize{
  \item Nikhil Gupta \email{guptan@smu.edu}
}

}
\keyword{internal}
