% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wn_tests.R
\name{white_noise_eval}
\alias{white_noise_eval}
\title{Evaluate if the dats is consistent with white noise}
\usage{
white_noise_eval(x, p = 0, q = 0, k_val = c(24, 48), alpha = 0.05,
  lag.max = 50)
}
\arguments{
\item{x}{the time series}

\item{p}{the ar order (Default = 0)}

\item{q}{the ma order (Default = 0)}

\item{k_val}{a vector of k values for ljung_box test}

\item{alpha}{Significance level to be used for ljung_box tests}

\item{lag.max}{Value of lags to plot for the ACF (Default: 50)}
}
\value{
the results of the tests, in tidy data format
}
\description{
Evaluate if the dats is consistent with white noise
}
\examples{
library(tswge)

# Generated White Noise 
wn = gen.arma.wge(n = 200, sn = 101)
white_noise_eval(wn)

# Not White Noise
data(hadley) 
white_noise_eval(hadley)

}
