% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicbic.R
\name{aicbic}
\alias{aicbic}
\title{aicbic function
returns the top 5 aic and bic values for an ARMA time series}
\usage{
aicbic(vec, ...)
}
\arguments{
\item{type}{either a string or plain text, defaults to "both", other values include AIC and BIC}
}
\description{
aicbic function
returns the top 5 aic and bic values for an ARMA time series
}
