% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{paramater estimation function}
\usage{
estimate(xs, p, q = 0, type = "mle", ...)
}
\arguments{
\item{xs}{the time series}

\item{p}{the ar order (defaults to 0)}

\item{type}{the method of estimation, defaults to maximum likelihood}
}
\value{
a list of estimates
}
\description{
paramater estimation function
}
