% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cor.R
\name{model_cor}
\alias{model_cor}
\title{Model Correlated data
Interactively build a model of correlated data}
\usage{
model_cor(xs, parallel = FALSE, cl = NULL)
}
\description{
Model Correlated data
Interactively build a model of correlated data
}
