% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCompareUnivariate.R
\name{ModelCompareUnivariate}
\alias{ModelCompareUnivariate}
\title{R6 class ModelCompareUnivariate}
\usage{
ModelCompareUnivariate$new(x = airlog, mdl_list = models,
                           n.ahead = 36, batch_size = 72)
}
\value{
A new `ModelCompareUnivariate` object.
}
\description{
Initialize an object to compare several Univatiate Time Series Models
}
\examples{
library(tswge)
data("airlog")

# Woodward Gray Airline Model
phi_wg = c(-0.36, -0.05, -0.14, -0.11, 0.04, 0.09, -0.02, 0.02, 0.17, 0.03, -0.10, -0.38)
d_wg = 1
s_wg = 12

# Parzen Model
phi_pz = c(0.74, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0.38, -0.2812)
s_pz = 12

# Box Model
d_bx = 1
s_bx = 12  
theta_bx =  c(0.40, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0.60, -0.24)

models = list("Woodward Gray Model A" = list(phi = phi_wg, d = d_wg, s = s_wg, sliding_ase = FALSE),
              "Woodward Gray Model B" = list(phi = phi_wg, d = d_wg, s = s_wg, sliding_ase = TRUE),
              "Parzen Model A" = list(phi = phi_pz, s = s_pz, sliding_ase = FALSE),
              "Parzen Model B" = list(phi = phi_pz, s = s_pz, sliding_ase = TRUE),
              "Box Model A" = list(theta = theta_bx, d = d_bx, s = s_bx, sliding_ase = FALSE),
              "Box Model B" = list(theta = theta_bx, d = d_bx, s = s_bx, sliding_ase = TRUE)
              )
              
mdl_compare = ModelCompareUnivariate$new(x = airlog, mdl_list = models,
                                         n.ahead = 36, batch_size = 72)
# Plots the historgam of the ASE values for each model.
# This is especially useful when models using a sliding window for ASE calculations. 
mdl_compare$plot_histogram_ases()

# The following method gives 2 plots
# (1) Plots the forecasts for each model along with the realization.
# (2) Plots the upper and lower limits for each model along with the realization.
# In both cases, this marks each batch using a background color for ease of comparison.
# only_sliding = TRUE will only plot forecsts for models using sliding ASE  calculations.
mdl_compare$plot_forecasts(only_sliding = TRUE)

# This method statistically compares all the models that use a sliding window ASE calculation
mdl_compare$statistical_compare()  

ASEs = mdl_compare$get_tabular_metrics(ases = TRUE)
print(ASEs)

# This method returns the metrics (ASE values) or forecasts for each model
# 'only_sliding' If set to TRUE, only the models that use a sliding window 
#                ASE calculation will be returned
# 'ases' If set to TRUE, this method will return the ASE value(s) 
#        Single value for models that don't use sliding ASEs and 
#        Multiple values (one per batch) for models that use sliding window
#        ASE calculations
#        If set to FALSE, this function will return the model forecasts and 
#        upper and lower confidence intervals. 
forecasts = mdl_compare$get_tabular_metrics(ases = FALSE)
print(forecasts)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelCompareUnivariate$new()}}
\item \href{#method-get_x}{\code{ModelCompareUnivariate$get_x()}}
\item \href{#method-get_batch_size}{\code{ModelCompareUnivariate$get_batch_size()}}
\item \href{#method-get_n.ahead}{\code{ModelCompareUnivariate$get_n.ahead()}}
\item \href{#method-add_models}{\code{ModelCompareUnivariate$add_models()}}
\item \href{#method-remove_models}{\code{ModelCompareUnivariate$remove_models()}}
\item \href{#method-compute_metrics}{\code{ModelCompareUnivariate$compute_metrics()}}
\item \href{#method-plot_histogram_ases}{\code{ModelCompareUnivariate$plot_histogram_ases()}}
\item \href{#method-plot_forecasts}{\code{ModelCompareUnivariate$plot_forecasts()}}
\item \href{#method-plot_simple_forecasts}{\code{ModelCompareUnivariate$plot_simple_forecasts()}}
\item \href{#method-plot_multiple_realizations}{\code{ModelCompareUnivariate$plot_multiple_realizations()}}
\item \href{#method-statistical_compare}{\code{ModelCompareUnivariate$statistical_compare()}}
\item \href{#method-get_tabular_metrics}{\code{ModelCompareUnivariate$get_tabular_metrics()}}
\item \href{#method-clone}{\code{ModelCompareUnivariate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$new(
  x = NA,
  mdl_list,
  n.ahead = NA,
  batch_size = NA,
  step_n.ahead = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Time Series Realization}

\item{\code{mdl_list}}{A names list of all models (see format below)}

\item{\code{n.ahead}}{The number of observations used to calculate ASE or forecast ahead}

\item{\code{batch_size}}{If any of the models used sliding ase method,
then this number indicates the batch size to use}

\item{\code{step_n.ahead}}{If using sliding window, should batches be incremented by n.ahead
(Default = TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_x"></a>}}
\if{latex}{\out{\hypertarget{method-get_x}{}}}
\subsection{Method \code{get_x()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$get_x()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_batch_size"></a>}}
\if{latex}{\out{\hypertarget{method-get_batch_size}{}}}
\subsection{Method \code{get_batch_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$get_batch_size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_n.ahead"></a>}}
\if{latex}{\out{\hypertarget{method-get_n.ahead}{}}}
\subsection{Method \code{get_n.ahead()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$get_n.ahead()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_models"></a>}}
\if{latex}{\out{\hypertarget{method-add_models}{}}}
\subsection{Method \code{add_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$add_models(mdl_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mdl_list}}{A names list of all models (see format below)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_models"></a>}}
\if{latex}{\out{\hypertarget{method-remove_models}{}}}
\subsection{Method \code{remove_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$remove_models(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Time Series Realization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-compute_metrics}{}}}
\subsection{Method \code{compute_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$compute_metrics(step_n.ahead = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_n.ahead}}{If using sliding window, should batches be incremented by n.ahead
(Default = TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_histogram_ases"></a>}}
\if{latex}{\out{\hypertarget{method-plot_histogram_ases}{}}}
\subsection{Method \code{plot_histogram_ases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$plot_histogram_ases()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_forecasts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_forecasts}{}}}
\subsection{Method \code{plot_forecasts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$plot_forecasts(only_sliding = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_simple_forecasts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_simple_forecasts}{}}}
\subsection{Method \code{plot_simple_forecasts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$plot_simple_forecasts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_multiple_realizations"></a>}}
\if{latex}{\out{\hypertarget{method-plot_multiple_realizations}{}}}
\subsection{Method \code{plot_multiple_realizations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$plot_multiple_realizations(
  n.realizations = 4,
  lag.max = 25,
  seed = NA,
  plot = c("all"),
  scales = "free_y"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-statistical_compare"></a>}}
\if{latex}{\out{\hypertarget{method-statistical_compare}{}}}
\subsection{Method \code{statistical_compare()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$statistical_compare()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tabular_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-get_tabular_metrics}{}}}
\subsection{Method \code{get_tabular_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$get_tabular_metrics(only_sliding = FALSE, ases = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelCompareUnivariate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
